# api/openapi.yaml
openapi: 3.0.0
info:
  title: API de Categorías
  version: 1.0.0
  description: Una API para gestionar categorías.
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
paths:
...
  /mercadopago/cashiers:
    get:
      summary: Obtiene todas las cajas
      operationId: getAllCashiers
      responses:
        '200':
          description: Lista de cajas (cashiers)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cashier'
        '400':
          description: Solicitud inválida (p.ej., mpStoreId faltante o inválido)
        '404':
          description: Tienda o cajas no encontradas
        '500':
          description: Error interno del servidor
    post:
      summary: Crea una nueva caja (cashier) en MercadoPago
      operationId: createCashier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                storeId:
                  type: string
                  description: ID de la tienda en MercadoPago a la que pertenece la caja
                cashierId:
                  type: number
                  description: ID de la caja en MercadoPago
                externalId:
                  type: string
                  description: ID externo para la caja
              required:
                - storeId
                - cashierId
                - externalId
      responses:
        '201':
          description: Caja creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cashier'
        '400':
          description: Solicitud inválida (p.ej., parámetros faltantes o inválidos)
        '500':
          description: Error interno del servidor
  /mercadopago/cashiers/{mpCashierId}:
    delete:
      summary: Elimina una caja específica por su ID
      operationId: deleteCashier
      parameters:
        - name: mpCashierId
          in: path
          required: true
          description: El ID de la caja de MercadoPago
          schema:
            type: string
      responses:
        '200':
          description: Caja eliminada exitosamente
        '400':
          description: Solicitud inválida (p.ej., mpCashierId faltante o inválido)
        '404':
          description: Caja no encontrada
        '500':
          description: Error interno del servidor
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    Cashier:
      type: object
      properties:
        id:
          type: number
          description: ID de la caja
        name:
          type: string
          description: Nombre de la caja
        dateCreation:
          type: string
          format: date-time
          description: Fecha de creación de la caja
        externalId:
          type: string
          nullable: true
          description: ID externo de la caja (si está disponible)
...